requirements.txt
Django==3.0.6
django-crontab==0.7.1
django-filter==2.2.0
djangorestframework==3.11.0
google-cloud-storage==1.28.1
Markdown==3.2.2
requests==2.23.0
